+++ lib/src/formatter/date_time_formatter.dart
@@ -1,0 +1,3 @@
+/// Simple formatting method for logs (up to seconds)
+String dateTimeUpToSeconds(DateTime time) =>
+    time.toIso8601String().replaceFirst('T', ' ').split('.').first;