+++ lib/vy_string_utils.dart
@@ -4,3 +4,5 @@ library vy_string_utils;
 
 export 'src/string_utils.dart';
 export 'src/string_extension.dart';
+
+export 'src/formatter/date_time_formatter.dart';